+++ package.json
@@ -88,7 +88,7 @@
         ],
         "devDependencies": {
                 "@axe-core/playwright": "4.4.5",
-                "@playwright/test": "1.29.2",
+                "@playwright/test": "1.26.1",
                 "@testing-library/react": "12.1.5",
                 "@testing-library/user-event": "14.4.3",
                 "@types/chromedriver": "81.0.1",
